created a wallet::send module and sorted send functionality into it
